+++ src/main/java/modelos/TransferenciaRecibida.java
@@ -16,7 +16,7 @@ public class TransferenciaRecibida extends Movimiento{
 	@ManyToOne
 	private Cuenta cuentaOrigen;
 
-	public TransferenciaRecibida(LocalDate fechayHora, float monto, String descripcion, Cuenta cuenta, Cuenta cuentaOrigen) {
+	public TransferenciaRecibida(LocalDate fechayHora, Double monto, String descripcion, Cuenta cuenta, Cuenta cuentaOrigen) {
 		super(fechayHora, monto, descripcion, cuenta);
 
 		this.cuentaOrigen=cuentaOrigen;
